+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add `UnsafeCellSlice::get_mut()`
+
+### Changed
+ - Support more index types in `UnsafeCellSlice::index_mut()`
+
 ## [0.2.0] - 2024-10-03
 
 ### Added